done: mvp is done

+++ Telegram.ApAzureBot.Core/Abstractions/Services/Web/Captcha/ICaptchaService.cs
@@ -2,5 +2,5 @@
 
 public interface ICaptchaService
 {
-    Task<ushort> SolveInteger(byte[] img, CancellationToken cToken);
+    Task<uint> SolveInteger(byte[] img, CancellationToken cToken);
 }
